app/requirements/base.txt
django==3.1
django-model-utils==4.0.0
py